doc: fix flaw in escape transactions.
As reported by mmeijeri on r/Bitcoin:

https://www.reddit.com/r/Bitcoin/comments/3dlxw4/reaching_the_ground_with_lightning_lightning/ct80xpp

As the signature doesn't cover the input itself, you could replace the
secret key from the escape transaction with 0, and it will use the "A and B"
branch of the output.

(This also fixes a bug in the anchor output where the OP_IF consumed the
secret).

Signed-off-by: Rusty Russell <rusty@rustcorp.com.au>