jsonrpc: Fail connect if TX broadcast fails
Rewired the broadcast_tx call to call the `broadcast_tx_complete`
callback once `bitcoin-cli` returns from `sendrawtransaction` so that we
can potentially report a failure if the broadcast failed. This also
added some more context to the `struct outgoing_tx` and `rawtx` which we
can potentially free after the broadcast completes.

Fixed #51