[EXP] Added payout to external address
Collaborative close now sends the coins to a configurable address, or
a P2PKH based on the peer_secrets saved in the database.

This means that providing `--payout-address` will result in the funds
being sent to that address. Not providing an address will result in a
P2PKH output, using the pubkey matching the connection's finalkey.