wire: Refactored common messages between daemons
There are a number of generic messages we'd like to ship around
between the main daemon and the subdaemon, independently of the daemon
type. For example we will want to be able to forward gossip messages
from and to the gossip daemon, without having special handling in the
main daemon. For now this is just the `status_trace` message and a
dummy message in the control.