subd: Added request muxing based on request numbers
So far we were always using the issue order and type to match replies
to requests. This has two problems: two concurrent requests with the
same type may get mixed up, and there is no way of streaming a
multipart response to the requestor, since the request object was
being freed after receiving the first reply.

This is an attempt at fixing both, by introducing explicit request
numbering and flags to signal the end of a reply.